+++ pages/index.js
@@ -9,23 +9,13 @@ import Title from '../components/Title';
 import TextRevealImageCard from '../components/TextRevealImageCard';
 import TextCard from '../components/TextCard';
 import Head from 'next/head'
-import ReactGA from 'react-ga'
-
-export const initGA = () => {
-  console.log('GA init')
-  ReactGA.initialize('UA-5819863-21')
-}
-export const logPageView = () => {
-  ReactGA.set({ page: window.location.pathname })
-  ReactGA.pageview(window.location.pathname)
-}
+import { logPageView } from '../utils/analytics'
 
 export default class extends React.Component {
 
   componentDidMount() {
     bgScrollEffect();
     initTabs();
-    initGA()
     logPageView()
   }
